+++ CHANGELOG.md
@@ -34,8 +34,9 @@
 
 ## [0.4.0] - 2016-08-14
 ### Added
-- Transforms and enum initializer from [`ember-flexberry`](https://github.com/Flexberry/ember-flexberry) addon.
+- Transforms, models, enums and enum initializer from [`ember-flexberry`](https://github.com/Flexberry/ember-flexberry) addon.
 - `decimal` transform for ability to use '.' and ',' as decimal separator in float numbers. So now it is possible to use `decimal` type for declaring attributes of models.
+- CRUD operations tests for OData.
 
 ### Changed
 - Function `enumCaptions` now returns object without null values.